@@ -5,22 +5,22 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.0
-  flame: ^0.27.0
-  bonsoir: ^0.1.2+2
-  synchronized: ^2.2.0+2
+  cupertino_icons: ^1.0.2
+  flame: ^1.0.0-rc7
+  flame_forge2d: ^0.6.2-rc4
+  flame_audio: ^0.1.0-rc4
+  flame_splash_screen: ^0.0.1
+  bonsoir: ^0.1.3+2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.0.0
-  build_test: ^0.10.3
 
 flutter:
 
