@@ -4,8 +4,9 @@ const PONG_PORT = 13579;
 const POP_FILE = 'pop.wav';
 const CRASH_FILE = 'crash.wav';
 const TADA_FILE = 'tada.wav';
-const WAHWAH_FILE = 'wahwah.wav';
-const OVERLAY_ID = 'Overlay';
+const WAH_FILE = 'wahwah.wav';
+const MAIN_MENU_OVERLAY_ID = 'MainMenuOverlay';
+const HOST_WAITING_OVERLAY_ID = 'HostWaitingOverlay';
 const MAX_SCORE = 3;
 const MARGIN_RATIO = .1; // of screen height
 const PAD_HEIGHT_RATIO = .01;
