@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mopong/game.dart';
+
+import '../lib/main.dart';
 
 void main() {
   testWidgets('MoPongGame test', (WidgetTester tester) async {
-    final game = MoPong();
-    runApp(game.widget);
+    final mopongApp = MoPongApp();
+    runApp(mopongApp);
   });
 }
